pyproject.toml
[tool.poetry]
name = "data-bench"
version = "0.1.0"
description = ""
authors = ["Aleksei Grachev <aleksei.grachev@alberblanc.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.7.2"
pandas = "2.1.4"
click = "^8.1.7"
human-readable = "^1.3.4"
dask = "^2024.2.1"
fastparquet = "^2024.2.0"
pyarrow = "^15.0.1"
jupyter = "^1.0.0"
python-dotenv = "^1.0.1"
tqdm = "^4.66.2"
ray = {version = "2.34.0", extras = ["default"]}
grpcio = "^1.62.1"
modin = "0.26.1"
torch = "2.4"
smart-open = {extras = ["s3"], version = "^7.0.4"}
millify = "^0.1.1"
setuptools = "69.5.1"
safetensors = "^0.5.2"
fsspec = "^2025.2.0"
s3fs = "^2025.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
